backend/requirements.txt
alembic==1.14.1
bcrypt==4.3.0
blinker==1.8.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.1
Faker==35.2.2
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
